feat: showMessage
