+++ apps/server/test/support/transport/no_op.ex
@@ -1,6 +1,5 @@
 defmodule Lexical.Test.Transport.NoOp do
   @behaviour Lexical.Server.Transport
 
-  def write(_message), do: :ok
-  def log(_level, _message), do: :ok
+  def write(_message, _opts \\ []), do: :ok
 end
